Technical challenge submission - v1.0

+++ backend/main.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI, Form
+from fastapi import FastAPI, Form, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 """
 This module defines a FastAPI application for managing a list of medicines.
@@ -32,16 +32,41 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-@app.get("/medicines")
-def get_all_meds():
+# Optimizing funtions to reduce code redundancy
+def read_data():
     """
-    This function reads the data.json file and returns all medicines.
+    This is a function that reads the data.json file and returns all medicines.
     Returns:
         dict: A dictionary of all medicines
     """
-    with open('data.json') as meds:
-        data = json.load(meds)
-    return data
+    try:
+        with open('data.json') as meds:
+            return json.load(meds)
+    except (FileNotFoundError, json.JSONDecodeError) as e:
+        raise HTTPException(status_code=500, detail="Error in reading the data")
+
+@app.get("/medicines")
+def get_all_meds():
+    """
+    This function returns all medicines from the data.json file.
+    Returns:
+        dict: A dictionary of all medicines
+    """
+    data = read_data()
+    medicines = []
+    for med in data["medicines"]:
+        if med['name'] not in (None, ""): # Handling missing values in the data file
+            if med['price'] not in (None, ""): # Handling null price values
+                medicines.append({
+                    "name": med['name'],
+                    "price": f"£{med['price']:.2f}" # Estimate all prices to 2 decimal points
+                })
+            else:
+                medicines.append({
+                    "name": med['name'],
+                    "price": "Price Not Available"
+                })
+    return {"medicines": medicines}
 
 @app.get("/medicines/{name}")
 def get_single_med(name: str):
@@ -52,13 +77,23 @@ def get_single_med(name: str):
     Returns:
         dict: A dictionary containing the medicine details
     """
-    with open('data.json') as meds:
-        data = json.load(meds)
-        for med in data["medicines"]:
-            print(med)
-            if med['name'] == name:
-                return med
-    return {"error": "Medicine not found"}
+    data = read_data()
+    for med in data["medicines"]:
+        if med['name'] == name:
+            if med['price'] not in (None, ""): # Handling null price values
+                return {
+                "name": med['name'],
+                "price": f"£{med['price']:.2f}"}
+            else:
+                return {
+                "name": med['name'],
+                "price": "Price not available"}
+
+    # Send error response if entered medicine name is not found
+    return {
+            "status": "error",
+            "code": "404",
+            "message": "Medicine not found"}
 
 @app.post("/create")
 def create_med(name: str = Form(...), price: float = Form(...)):
@@ -71,13 +106,23 @@ def create_med(name: str = Form(...), price: float = Form(...)):
     Returns:
         dict: A message confirming the medicine was created successfully.
     """
-    with open('data.json', 'r+') as meds:
-        current_db = json.load(meds)
-        new_med = {"name": name, "price": price}
-        current_db["medicines"].append(new_med)
-        meds.seek(0)
-        json.dump(current_db, meds)
-        meds.truncate()
+    # Send error response if user tries to create entries with null/empty values
+    if name in (None, "") or price in (None, ""):
+        return {
+            "status": "error",
+            "code": "400",
+            "message": "Provide valid entries for Name and Price"}
+
+    try:
+        with open('data.json', 'r+') as meds:
+            current_db = json.load(meds)
+            new_med = {"name": name, "price": price}
+            current_db["medicines"].append(new_med)
+            meds.seek(0)
+            json.dump(current_db, meds)
+            meds.truncate() # Problem: Could cause issues during concurrent writes.
+    except (FileNotFoundError, json.JSONDecodeError) as e:
+        raise HTTPException(status_code=500, detail="Error in reading the data")
         
     return {"message": f"Medicine created successfully with name: {name}"}
 
@@ -92,16 +137,32 @@ def update_med(name: str = Form(...), price: float = Form(...)):
     Returns:
         dict: A message confirming the medicine was updated successfully.
     """
-    with open('data.json', 'r+') as meds:
-        current_db = json.load(meds)
-        for med in current_db["medicines"]:
-            if med['name'] == name:
-                med['price'] = price
-                meds.seek(0)
-                json.dump(current_db, meds)
-                meds.truncate()
-                return {"message": f"Medicine updated successfully with name: {name}"}
-    return {"error": "Medicine not found"}
+
+    if name in (None, "") or price in (None, ""):
+        return {
+            "status": "error",
+            "code": "400",
+            "message": "Provide valid entries for Name and Price"
+        }
+
+    try:
+        with open('data.json', 'r+') as meds:
+            current_db = json.load(meds)
+            for med in current_db["medicines"]:
+                if med['name'] == name:
+                    med['price'] = price
+                    meds.seek(0)
+                    json.dump(current_db, meds)
+                    meds.truncate()
+                    return {"message": f"Medicine updated successfully with name: {name}"}
+    except (FileNotFoundError, json.JSONDecodeError) as e:
+        raise HTTPException(status_code=500, detail="Error in reading the data")
+    # Send error response if entered medicine name is not found
+    return {
+        "status": "error",
+        "code": "404",
+        "message": "Medicine not found"
+    }
 
 @app.delete("/delete")
 def delete_med(name: str = Form(...)):
@@ -113,18 +174,41 @@ def delete_med(name: str = Form(...)):
     Returns:
         dict: A message confirming the medicine was deleted successfully.
     """
-    with open('data.json', 'r+') as meds:
-        current_db = json.load(meds)
-        for med in current_db["medicines"]:
-            if med['name'] == name:
-                current_db["medicines"].remove(med)
-                meds.seek(0)
-                json.dump(current_db, meds)
-                meds.truncate()
-                return {"message": f"Medicine deleted successfully with name: {name}"}
-    return {"error": "Medicine not found"}
+    try:
+        with open('data.json', 'r+') as meds:
+            current_db = json.load(meds)
+            for med in current_db["medicines"]:
+                if med['name'] == name:
+                    current_db["medicines"].remove(med)
+                    meds.seek(0)
+                    json.dump(current_db, meds)
+                    meds.truncate()
+                    return {"message": f"Medicine deleted successfully with name: {name}"}
+    except (FileNotFoundError, json.JSONDecodeError) as e:
+        raise HTTPException(status_code=500, detail="Error in reading the data")
+    # Send error response if entered medicine name is not found
+    return {
+            "status": "error",
+            "code": "404",
+            "message": "Medicine not found"}
 
 # Add your average function here
+@app.get("/average_price")
+def get_average_price():
+    """
+    This function calculate and returns the average price of all medicines.
+    """
+    data = read_data()
+    total_price = 0
+    count = 0
+    for med in data["medicines"]:
+        if med['price'] not in (None, ""):
+            total_price += med['price']
+            count += 1
+    if count == 0:
+        return {"average_price": "No valid prices available"}
+    average_price = total_price / count
+    return {"average_price": f"£{average_price:.2f}"}
 
 if __name__ == "__main__":
     uvicorn.run(app, host="0.0.0.0", port=8000)
